+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 Bugfixes:
 * Fixes `debug` log messages not being printed
 * Pressing Ctrl+C wouldn't stop package identification with KeyboardInterrupt
+* Fixes an error message while destroying a docker experiment
 
 0.5.1 (2014-12-18)
 ------------------